% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_dynamics.R
\name{check_assay_order}
\alias{check_assay_order}
\title{Check that assays are ordered by window period}
\usage{
check_assay_order(
  list_of_assays,
  short_window_period_first = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{list_of_assays}{A character vector specifying the names of the assays. The order of this list will be checked.}

\item{short_window_period_first}{If True, check that the vector is ordered from the assay with the shortest window period to the assay with the longest window period.}

\item{verbose}{If the order is incorrect, a warning will be raised specifying each sequential pair that is incorrectly ordered.}
}
\description{
Given a list (format: character vector of names) of assays, compare pairs of sequential assays using which_is_faster to ensure that they are ordered correctly. If the order is correct, then TRUE will be returned. If not, then FALSE will be returned and a warning will be raised if verbose = TRUE.
}
