% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_processing.R
\name{check_lb_med_ub}
\alias{check_lb_med_ub}
\title{Basic function used for testing estimate_lb_med_ub}
\usage{
check_lb_med_ub(lb, med, ub, fun, range_start, range_end)
}
\description{
Given a function and the lb, med and ub, verify that the lb, med and ub are correct.
}
