% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{combine_sc_and_visit_times}
\alias{combine_sc_and_visit_times}
\title{Produce ihist from sc and visit times}
\usage{
combine_sc_and_visit_times(sc_times, visit_times,
  true_infection_date = 17000.5, ptid = "?")
}
\description{
Given a list of sc times as produced by sim_sc_times, and a vector of visit times relative to the time of true infection (t = 0 is the time of true infection) produce an ihist. This is done by comparing the visit time with the seroconversion times and deducing the result for the relevant test at the visit. Lastly, the dates are converted so that the true infection time is at an appropriate calendar date and not t = 0.
}
\details{
TODO: Should there be an option that places the first positive visit at t = 0?
}
