% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_processing.R
\name{compute_daily_grid}
\alias{compute_daily_grid}
\title{Computes probability that infection was on each of a range of days}
\usage{
compute_daily_grid(agg_fun, tauc, range_start, range_end)
}
\arguments{
\item{agg_fun}{The aggregate curve associated with the ihist of interest as computed by construct_aggregate_interpreter.}

\item{tauc}{The total area under the aggregate curve.}

\item{range_start}{The start of the range over which daily probabilities should be computed.}

\item{range_end}{The end of the range over which daily probabilities should be computed.}
}
\description{
Given an aggregate curve, a range of dates and the total area under the curve compute a vector of probabilities so that each element reflects the probability that infection occurred on a specific day.
}
