% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_interpretation.R
\name{construct_assay_result_interpreter}
\alias{construct_assay_result_interpreter}
\title{Construct an assay result interpretation function}
\usage{
construct_assay_result_interpreter(assay_dynamics, result)
}
\arguments{
\item{assay_dynamics}{A list providing the assay_dynamics, (TODO: elaborate)}

\item{result}{The result of the test. Either '+' or '-'.}
}
\description{
Given the dynamics of an assay and a result, produce a function that when evaluated for a given number of days between the exposure event and the day the sample was drawn will produce the probability of observing the given result. This function will only take x as input, which is the number of days, and will already have the particulars of the result and assay loaded into it.
}
