% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_processing.R
\name{convert_aggregate_into_ecdf}
\alias{convert_aggregate_into_ecdf}
\title{Convert Aggregate prob curve into an ecdf.}
\usage{
convert_aggregate_into_ecdf(dat)
}
\arguments{
\item{dat}{An interpreted resultset in long format}
}
\description{
Convert Aggregate prob curve into an ecdf.
}
