% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_processing.R
\name{convert_interpretation_to_long}
\alias{convert_interpretation_to_long}
\title{Converts an interpretated results set to long format}
\usage{
convert_interpretation_to_long(interpretation_set)
}
\arguments{
\item{interpretation_set}{An interpretated results set as produced by interpret_result_set}
}
\description{
Converts an interpretated result set as produced by interpret_result_set to long format
}
