% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_weibull3.R
\name{fit_weibull3}
\alias{fit_weibull3}
\title{Fits a 3 parameter Weibull distribution to percentiles}
\usage{
fit_weibull3(x, y)
}
\arguments{
\item{x}{The percentiles. The actual value that correspond to the y'th percentile.}

\item{y}{The percentile values (between 0 and 1)}
}
\description{
The Delaney paper reports the 25th, 50th, 75th and 99th percentiles of the window periods of various assays. This function will fit a three parameter Weibull distribution to these four points so that it can be included in the library.
}
