% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_dynamics.R
\name{get_assay_dynamics}
\alias{get_assay_dynamics}
\title{Get dynamics functions for an assay}
\usage{
get_assay_dynamics(assay)
}
\arguments{
\item{assay}{The name of the assay whose dynamics are required.}
}
\description{
Constructs the assay dynamics list with suitable parameter set given the name of an assay
}
