% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_interpretation.R
\name{get_scatterpoints}
\alias{get_scatterpoints}
\title{Evaluates a function over a range}
\usage{
get_scatterpoints(fun, seedpoints)
}
\arguments{
\item{fun}{The function to evaluate.}

\item{seedpoints}{The points at which the function should be evaluated initially. It is very important to provide the function with a sensible set of seedpoints. Due to the design of the tsic package, such a set of points should always be available.}
}
\description{
Given a function and either the endpoints of a range of a number of seedpoints, evaluate the function at points in the range so that the change in y does not exceed a certain threshold between any two points, subject to the constraint that no interval between to x points go below some interval.
}
\details{
Note that
}
