% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_interpretation.R
\name{interpret_ihist}
\alias{interpret_ihist}
\title{Interprets an ihist into daily likelihoods}
\usage{
interpret_ihist(ihist, range_start, range_end, verbose = FALSE)
}
\arguments{
\item{ihist}{The diagnostic history for an individual.}

\item{range_start}{The earliest day of interest.}

\item{range_end}{The latest day of interest.}

\item{verbose}{Should verbose output be printed during the computation?}
}
\description{
Given a range of interest and an individual's diagnostic history, compute the likelihood that each day in the range was the day of infection. This produces the output in long format.
}
