% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_interpretation.R
\name{interpret_result}
\alias{interpret_result}
\title{Probability of DDI_1 curve implied by test result}
\usage{
interpret_result(assay_dynamics, result, range_start, range_end,
  visit_date, min_prob = 0, max_prob = 1)
}
\arguments{
\item{assay_dynamics}{A list providing the assay_dynamics, (TODO: elaborate)}

\item{result}{'+' or '-' indicating the test result.}

\item{range_start}{The date furthest into the past that should be considered as a plausible day for DDI_1.}

\item{range_end}{The most recent date that should be considered as a plausible day for DDI_1.}

\item{visit_date}{The date of the visit on whose sample the test result was observed}

\item{min_prob}{The most 'certain' you are that the given test results excludes dates far from the actual visit. Think of this as a chance that the test may give an incorrect result.}

\item{max_prob}{The most 'certain' you are that the given test results excludes dates far from the actual visit. Think of this as a chance that the test may give an incorrect result. TODO: Think more about what this actually means.}
}
\description{
Compute a curve of probabilities of observing a specified test result on a specified date assuming the day of DDI_1. Thus assume individually that each day in a range of dates is the DDI_1 and then compute the probability of observing the specified result.
}
