% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_of_diagnostics.R
\name{interpret_result_set}
\alias{interpret_result_set}
\title{Given a set of results, construct DDI_1 probability curves}
\usage{
interpret_result_set(result_set, range_start = NULL, range_end = NULL)
}
\arguments{
\item{result_set}{A data.frame of all the test results and dates for a patient}

\item{range_start}{The date furthest into the past that should be considered as a plausible day for DDI_1.}

\item{range_end}{The most recent date that should be considered as a plausible day for DDI_1.}
}
\description{
Calls interpret_result on a set of test results and aggregate the results into a data.frame
}
