% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dsmb_nov_2019_data}
\alias{load_dsmb_nov_2019_data}
\title{Loads dataset used for DSMB Nov 2019}
\usage{
load_dsmb_nov_2019_data(file_name)
}
\arguments{
\item{file_name}{Full path to the csv file}
}
\description{
Loads dataset used for DSMB Nov 2019
}
