% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_presentation.R
\name{make_vlines_dat}
\alias{make_vlines_dat}
\title{Create Dataset of test dates}
\usage{
make_vlines_dat(in_dat)
}
\arguments{
\item{in_dat}{An interpretated datasets in long format}
}
\description{
Prepares a dataset with the dates of the tests for easy inclusion in plots.
}
