% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_processing.R
\name{manual_approx}
\alias{manual_approx}
\title{Interpolates - multiple points}
\usage{
manual_approx(x, y, xout)
}
\arguments{
\item{x, }{y vectors giving the coordinates of the points to be interpolated.}

\item{xout}{Vector of values specifying where interpolation is to take place.}
}
\description{
Crappy function needed because internal approx fails
}
