% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_processing.R
\name{manual_approx_one}
\alias{manual_approx_one}
\title{Interpolates - single point}
\usage{
manual_approx_one(x, y, xout)
}
\arguments{
\item{x, }{y vectors giving the coordinates of the points to be interpolated.}

\item{xout}{Single value specifying where interpolation is to take place.}
}
\description{
Crappy function needed because internal approx fails
}
