% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{parse_data_first_mock}
\alias{parse_data_first_mock}
\title{Parse and prepare the data for the first mock dataset}
\usage{
parse_data_first_mock(file_name)
}
\arguments{
\item{file_name}{Full path the the csv file}
}
\description{
This script will load the input file and prepare it for processing. It is currently specifically designed to be used with the mock datasets. A more general version of this function must be written.
}
