% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_presentation.R
\name{patient_plot_compact}
\alias{patient_plot_compact}
\title{Plots a patients timelines compactly}
\usage{
patient_plot_compact(lrs, vlines)
}
\arguments{
\item{lrs}{An interpretated result set in long format.}

\item{vlines}{Vertical lines indicating the test dates as produced by make_vlines_dat.}
}
\description{
Plots dates and the likelihoods that these dates are the DDI_1 for a patient.
}
