% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_presentation.R
\name{plot_iihist}
\alias{plot_iihist}
\title{Plots an individual history}
\usage{
plot_iihist(ihist, lb_med_ub, range_start, range_end,
  produce_plot = TRUE, save_plot = FALSE, verbose = FALSE,
  plot_aggregate = TRUE)
}
\description{
Will produce the interpretation itself.
}
