% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_of_diagnostics.R
\name{prob_test_result_given_days_to_visit}
\alias{prob_test_result_given_days_to_visit}
\title{Probabilities of certain test results}
\usage{
prob_test_result_given_days_to_visit(assay_dynamics, result, days_to_visit)
}
\arguments{
\item{assay_dynamics}{A list providing the assay_dynamics, (TODO: elaborate)}

\item{result}{'+' or '-' indicating the test result.}

\item{days_to_visit}{The number of days between DDI_1 and the test result.}
}
\description{
Turns assay dynamics into the probability of observing a specified test result a given number of days since DDI_1
}
