% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{recognize_scenario}
\alias{recognize_scenario}
\title{Recognize a scenario from an ihist}
\usage{
recognize_scenario(ihist, scenario_mapping = NULL)
}
\arguments{
\item{ihist}{The ihist from which the scenario must be detected.}

\item{scenario_mapping}{A list of lists where the names in the outer list names the scenarios and the inner lists provides the test patterns that are associated with that scenario. If no scenario_mapping is specified, the AMP Taqman/Architect mapping will be used.}
}
\description{
Given a scenario mapping and an ihist, this function will detect what scenario the given ihist represents.
}
\details{
Note that this function will apply the 'select_most_informative_results' function to the ihist, so that places some limits on how exotic the scenario description can be.

TODO: This function will be much more useful if it is extended to handle different assays that may produce the same result. For example, allowing the scenario 1 case to be defined by a negative on with Taqman or Abbott Real Time.
}
