% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_processing.R
\name{remove_non_informative_results}
\alias{remove_non_informative_results}
\title{Remove non-informative test results}
\usage{
remove_non_informative_results(in_dat)
}
\arguments{
\item{in_dat}{An interpreted resultset in long format}
}
\description{
For each assay, if this result is negative and the next one is also negative, remove this result. For each assys, this this result is positive and the previous result is also positive, then remove this result. Obviously this needs refinement.
}
