% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{remove_strongly_dependent_results}
\alias{remove_strongly_dependent_results}
\title{Removes strongly dependent results at the same timepoint}
\usage{
remove_strongly_dependent_results(
  ihist,
  more_sensitive_test,
  less_sensitive_test
)
}
\arguments{
\item{ihist}{The diagnostic history from which to remove such cases}

\item{more_sensitive_test}{The test which detects infection faster. In case both are positive, this one will be removed.}

\item{less_sensitive_test}{The test which detects infection slower. In case both are negative, this one will be removed.}
}
\description{
The construction of the aggregate curve assumes that the contributions from the different tests are independent. For some tests, for example fully reactive and indeterminate versions of the same test this assumption is strongly violated. To offset this, remove the least informative version of the in case that they give the same result.
}
