% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{select_most_informative_results}
\alias{select_most_informative_results}
\title{Selects most informative test results only}
\usage{
select_most_informative_results(ihist, fastest_to_slowest_tests = NULL)
}
\arguments{
\item{ihist}{The diagnostic test history of the individual.}

\item{fastest_to_slowest_tests}{A vector listing the tests in the order from the test with the shortest window period to the test with the longest window period. This ordering will be consulted in conjunction with the test result to decide which tests are the most informative.}
}
\description{
Given a diagnostic history and an ordered list of tests by window period, this function will return a restricted diagnostic history that only contains the most informative test results for each sample date. The most informative results are the test with the shortest window period that produced a negative result and the test with the longest window period that produced a positive result. The motivation for doing this is to ensure that the aggregate function is constructed using independent results.
}
