% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_durations}
\alias{sim_durations}
\title{Simulates the durations times between fiebig stages}
\usage{
sim_durations(n = 30, rates = "fiebig",
  eclipse_distribution = list(distr = function(n, shape, scale, location)
  {     location + rweibull(n, shape = shape, scale = scale) }, params =
  list(shape = 1.35, scale = 9, location = 4.8)))
}
\arguments{
\item{n}{The number of people to simulate}

\item{rates}{The parameters for the exponential distributions of each of the six stages. ( default = 'fiebig' which gets replaced with list(l1 = 5.0, l2 = 5.3, l3 = 3.2, l4 = 5.6, l5 = 69.5) )}

\item{eclipse_distribution}{A list with elements distr pointing to the function allowing draws from the eclipse distribution ( default = function(n, shape, scale, location) {location + rweibull(n, shape = shape, scale = scale)} ) and the parameters of said distribution ( default = list(shape = 1.35, scale = 9, location = 4.8) )}
}
\value{
data.frame with columns for ptid and the drations of the eclipse phase and stages I to V.
}
\description{
For each patients, draw an eclispe duration from the supplied eclipse_distribution,
and stage durations from exponential distributions with parameters specified by the rates argument.
}
\details{
Only the durations in stages I to V is drawn since the exit form stage VI is not defined / measured.
}
