% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_dx_results}
\alias{sim_dx_results}
\title{Simulates a visit's diagnostic results}
\usage{
sim_dx_results(tsi, list_of_assays)
}
\arguments{
\item{tsi}{The time since infection in days. tsi = 0 is the day of infection.}

\item{list_of_assays}{ORDERED list of assays for which test results should be produced. They must be ordered with the fastest assay first (fastest = assay with the shortest window period)}
}
\description{
Given the time since infection, this function will draw from the window period distributions of the specified assays to obtain a set of diagnostic test result.
}
\details{
The first negative test result will stop the procedure since then it is known that all other assays will also produce negative results.
}
