% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_test_results}
\alias{sim_test_results}
\title{Reads test results from a duration and visit schedule}
\usage{
sim_test_results(durations, visit_schedule,
  tests_and_properties = list(aptima_weib3_delaney = list(name =
  aptima_weib3_delaney, stage = 1, offset = -2), architech_weib3_delaney =
  list(name = architech_weib3_delaney, stage = 2, offset = -1),
  geenius_indet_weib3_delaney = list(name = geenius_indet_weib3_delaney,
  stage = 3, offset = 0), geenius_fr_weib3_delaney = list(name =
  geenius_fr_weib3_delaney, stage = 5, offset = 0)))
}
\arguments{
\item{durations}{The data.frame of transitions as produced by sim_durations.}

\item{visit_schedule}{The data.frame with the date of each visit as produced by sim_visit_dates.}

\item{tests_and_properties}{A list with each element corresponding to a test giving the Fiebig stage in which the test first starts returning positive results. Additionally, each test has an offset that allows the user to specify that the test starts testing positive x days before or after person enters the stage.}
}
\description{
Reads test results from a duration and visit schedule
}
