% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_visit_dates}
\alias{sim_visit_dates}
\title{Simulates the visit dates around a set of transitions}
\usage{
sim_visit_dates(durations, inf_date = as.numeric(as.Date("2015-05-01")),
  n_before_infection = 2, gap_distribution = function(n) {     rnorm(n,
  30, 3) })
}
\arguments{
\item{durations}{The data.frame of transitions as produced by sim_durations.}

\item{inf_date}{The infection date. In this version, all infections occur on the same date. This makes comparisions easy also.}

\item{n_before_infection}{The number of visits before infection.}

\item{gap_distribution}{A function that generates the n gaps between the next n visits.}
}
\description{
Simulates the visit dates around a set of transitions
}
