% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{visit_labeller}
\alias{visit_labeller}
\title{Annotates an ihist with visit_numbers and rel_sample_dates}
\usage{
visit_labeller(ihist)
}
\arguments{
\item{ihist}{An ihist.}
}
\description{
Given an ihist with the four main columns, add a visit_number column with the first visit at which any test was positive as 0. The visit immediately before the first positive visit is visit_number -1 while the visit immediately after the first positive visit is visit_number 1.
}
