% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_dynamics.R
\name{weib3_assay_dynamics}
\alias{weib3_assay_dynamics}
\title{Weibull3 Assay Dynamics}
\usage{
weib3_assay_dynamics(x, location, scale, shape)
}
\arguments{
\item{x}{The time since infection}

\item{location}{The location parameter.}

\item{scale}{The scale parameter.}

\item{shape}{The shape parameter.}
}
\description{
Returns the probability of testing positive x days after infection
}
\details{
Assumes that the probability of testing positive as a function of time given that the person is infected has the form of a Weibull distribution.
}
