% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_interpretation.R
\name{which_is_faster}
\alias{which_is_faster}
\title{Given two assays, returns which is faster}
\usage{
which_is_faster(assay1, assay2, comp_range = (8000:10005)/10)
}
\arguments{
\item{assay1}{The first assay. Needs to be the full assay characterization as returned by get_assay_dynamics}

\item{assay2}{The second assay. Needs to be the full assay characterization as returned by get_assay_dynamics}

\item{comp_range}{A vector of all the timepoints at which the two assays should be compared. The comparison will be on the basis of a positive test at the highest number in this vector. NOTE: the points in this vector must be evenly spaced.}
}
\description{
A crude function that computes the areas between the curves. If the area bounded on the top by assay1 is larger than the area bounded on the top by assay2, then assay1 is said to be the faster assay.
}
