% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_presentation.R
\name{wrapped_patient_plot}
\alias{wrapped_patient_plot}
\title{Plots and annotates a patients timelines}
\usage{
wrapped_patient_plot(dat)
}
\arguments{
\item{dat}{An interpretated result set in long format.}
}
\description{
Plots the dates and the likelihoods that these dates are the DDI_1 for a
patient. Plots are annotated with the test dates as vertical lines.
}
