% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_interpretation.R
\name{construct_aggregate_interpreter}
\alias{construct_aggregate_interpreter}
\title{Constructs an interpreter for a diagnostic history}
\usage{
construct_aggregate_interpreter(ihist)
}
\arguments{
\item{ihist}{A data.frame with columns sample_date, test, and result that contains a diagnostic history (or a subset thereof) for a single person.}
}
\description{
Construct a closure that gives the likelihood of a diagnostic history if the infection event was on the date supplied.
}
