% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_interpretation.R
\name{construct_assay_result_interpreter}
\alias{construct_assay_result_interpreter}
\title{Construct an assay result interpretation function}
\usage{
construct_assay_result_interpreter(assay_dynamics, result, sample_date)
}
\arguments{
\item{assay_dynamics}{A list providing the assay_dynamics, (TODO: elaborate)}

\item{result}{The result of the test. Either '+' or '-'.}

\item{sample_date}{The date (in days since 1970-01-01) on which the sample was drawn.}
}
\description{
Given the dynamics of an assay and a result with a date (in days since 1970-01-01), produce a function that when evaluated for a potential exposure date (in days since 1970-01-01) will produce the probability of observing the given result at the relevant sample date. This function will only take x as input, which is the potential exposure date (in days since 1970-01-01), and will already have the particulars of the result and assay loaded into it.
}
