% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_processing.R
\name{estimate_lb_med_ub}
\alias{estimate_lb_med_ub}
\title{Estimate the LB, UB and median for DDI_1}
\usage{
estimate_lb_med_ub(dat)
}
\arguments{
\item{dat}{An interpreted resultset in long format}
}
\description{
Estimate the LB, UB and median for DDI_1
}
