% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_processing.R
\name{estimate_lb_med_ub}
\alias{estimate_lb_med_ub}
\title{Estimates the 2.5, 50 and 97.5 percentiles}
\usage{
estimate_lb_med_ub(fun, range_start, range_end, verbose = FALSE,
  label = "unlabeled")
}
\arguments{
\item{fun}{The function whose percentiles are required}

\item{range_start}{Start of interval containing the percentiles}

\item{range_end}{End of interval containing the percentiles}

\item{verbose}{Should verbose output be printed?}

\item{label}{A label to print out with warnings. ptid is a good candidate}
}
\description{
Various hacks required to work around the inadequacy of R's numerical analysis tools.
}
