% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_processing.R
\name{estimate_lb_med_ub}
\alias{estimate_lb_med_ub}
\title{Estimates the 2.5, 50 and 97.5 percentiles and much more}
\usage{
estimate_lb_med_ub(
  fun,
  range_start,
  range_end,
  verbose = FALSE,
  label = "unlabeled",
  warn_low_AOC = FALSE,
  extra_tiles = NULL,
  date_splits = NULL
)
}
\arguments{
\item{fun}{The function whose percentiles are required}

\item{range_start}{Start of interval containing the percentiles}

\item{range_end}{End of interval containing the percentiles}

\item{verbose}{Should verbose output be printed?}

\item{label}{A label to print out with warnings. ptid is a good candidate}

\item{warn_low_AOC}{Should a warning be issued when AOC is low? default = FALSE}

\item{extra_tiles}{A vector of additional percentiles to compute. If NULL, this process will be skipped.}

\item{date_splits}{Compute the AOC to the left of each of the dates listed in this parameter. If NULL, this process will be skipped.}
}
\description{
Various hacks required to get the numerical integration and optimization working.
}
