% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_dynamics.R
\name{get_assay_dynamics}
\alias{get_assay_dynamics}
\title{Get dynamics functions for an assay}
\usage{
get_assay_dynamics(assay = NULL)
}
\arguments{
\item{assay}{The name of the assay whose dynamics are required. If NULL (default), list of available assays will be returned. If 'all' a list of all assay dynamics will be returned.}
}
\description{
Constructs the assay dynamics list with suitable parameter set given the name of an assay
}
