% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_interpretation.R
\name{interpret_ihist}
\alias{interpret_ihist}
\title{Interprets an ihist into daily likelihoods}
\usage{
interpret_ihist(ihist, range_start, range_end, verbose = FALSE)
}
\arguments{
\item{ihist}{The diagnostic history for an individual.}

\item{range_start}{The earliest day of interest.}

\item{range_end}{The latest day of interest.}

\item{verbose}{Should verbose output be printed during the computation?}
}
\description{
Given a range of interest and an individual's diagnostic history, compute the likelihood that each day in the range was the day of infection. This produces the output in long format.
}
\examples{
ihist <- data.frame(
  ptid = c('p0', 'p0'),
  sample_date = c('2016-03-01', '2016-09-01'),
  test = c('step_unit_testing', 'step_unit_testing'),
  result = c('-', '+'),
  stringsAsFactors = FALSE
)
range_start <- as.Date('2016-01-01')
range_end <- as.Date('2016-12-01')
interpret_ihist(ihist = ihist, range_start = range_start, range_end = range_end, verbose = TRUE)
}
